+++ test/modelValidatorTests.ts
@@ -607,4 +607,21 @@ describe("Model Validation", () => {
       assert.strictEqual(result.length, 1);
     });
   });
+
+  describe("should pass for validate generated 'uri' format string", () => {
+    it("should pass when examples match the 'uri' format definition of swagger file", async () => {
+      const specPath2 = `${testPath}/modelValidation/swaggers/specification/loadtestservice/loadtestservice.json`;
+      const result = await validate.validateExamples(specPath2, 'TestRun_StopTestRun');
+      assert.strictEqual(result.length, 0);
+    });
+
+    it("should failed when examples doesn't match the 'uri' format definition of swagger file", async () => {
+      const specPath2 = `${testPath}/modelValidation/swaggers/specification/loadtestservice/loadtestservice.json`;
+      const result = await validate.validateExamples(specPath2, 'TestRun_GetAppTestRunsSearch');
+      assert.strictEqual(result.length, 3);
+      assert.strictEqual(result[0].code, "INVALID_FORMAT");
+      assert.strictEqual(result[1].code, "INVALID_FORMAT");
+      assert.strictEqual(result[2].code, "INVALID_FORMAT");
+    });
+  });
 });