+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - Traffic validation - support multiple swaggers and multiple payloads
 - Extends format types in AJV schema validation
 - Bug fix - check discriminator property type must be string
+- Bug fix for generate api scenario
 - Upgrade autorest scheams about x-ms-enum change
 - Model/Live validation - fix content-type validation and unknown format issue during validator building